Revision errores, Habitacion

+++ src/main/java/ar/edu/utn/frmdp/ultimate_hotel_software/service/serviceImpl/HabitacionServiceImpl.java
@@ -2,17 +2,17 @@ package ar.edu.utn.frmdp.ultimate_hotel_software.service.serviceImpl;
 
 
 import ar.edu.utn.frmdp.ultimate_hotel_software.enums.EstadoHabitacion;
-import ar.edu.utn.frmdp.ultimate_hotel_software.exception.HabitacionDuplicadaException;
-import ar.edu.utn.frmdp.ultimate_hotel_software.exception.HabitacionNoEncontradaException;
-import ar.edu.utn.frmdp.ultimate_hotel_software.exception.InvalidIdException;
+import ar.edu.utn.frmdp.ultimate_hotel_software.exception.*;
 import ar.edu.utn.frmdp.ultimate_hotel_software.mapper.HabitacionMapper;
 import ar.edu.utn.frmdp.ultimate_hotel_software.models.HabitacionEntity;
 import ar.edu.utn.frmdp.ultimate_hotel_software.models.dto.requests.HabitacionDTORequest;
 import ar.edu.utn.frmdp.ultimate_hotel_software.models.dto.requests.HabitacionUpdateDTO;
+import ar.edu.utn.frmdp.ultimate_hotel_software.models.dto.response.EmpleadoDTOResponse;
 import ar.edu.utn.frmdp.ultimate_hotel_software.models.dto.response.HabitacionDTOResponse;
 import ar.edu.utn.frmdp.ultimate_hotel_software.repository.HabitacionRepository;
 import ar.edu.utn.frmdp.ultimate_hotel_software.service.HabitacionService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -24,6 +24,44 @@ public class HabitacionServiceImpl implements HabitacionService {
     private final HabitacionRepository habitacionRepository;
     private final HabitacionMapper habitacionMapper;
 
+    //1. Busquedas por ID
+    //1.1. Devuelve entidad
+    @Override
+    public HabitacionEntity findEntityById(Long id) {
+
+        return habitacionRepository.findById(id)
+                .orElseThrow(() ->
+                        new HabitacionNoEncontradaException("Habitación no encontrada"));
+    }
+
+    //1.2. Devuelve DTO Response
+    @Override
+    public HabitacionDTOResponse findById(Long id) {
+
+        HabitacionEntity habitacion = habitacionRepository.findById(id)
+                .orElseThrow(() ->
+                        new HabitacionNoEncontradaException("Habitación no encontrada"));
+
+        return habitacionMapper.toResponse(habitacion);
+    }
+
+    //2. Listar habitaciones
+    //2.1. Listar entidades
+    @Override
+    public List<HabitacionEntity> findAll() {
+        return habitacionRepository.findAll();
+    }
+
+    //2.2. Listar DTOs Response
+    @Override
+    public List<HabitacionDTOResponse> getAll() {
+        return habitacionRepository.findAll().stream()
+                .map(habitacionMapper::toResponse)
+                .toList();
+    }
+
+
+    //3. Crear habitacion
     @Override
     public HabitacionDTOResponse save(HabitacionDTORequest dto) {
 
@@ -40,26 +78,25 @@ public class HabitacionServiceImpl implements HabitacionService {
         return habitacionMapper.toResponse(guardada);
     }
 
-    @Override
-    public HabitacionDTOResponse findById(Long id) {
 
-        HabitacionEntity habitacion = habitacionRepository.findById(id)
-                .orElseThrow(() ->
-                        new HabitacionNoEncontradaException("Habitación no encontrada"));
-
-        return habitacionMapper.toResponse(habitacion);
-    }
-
+    //4. Borrar habitacion
     @Override
     public void delete(Long id) {
 
         HabitacionEntity habitacion = habitacionRepository.findById(id)
                 .orElseThrow(() ->
                         new HabitacionNoEncontradaException("Habitación no encontrada"));
+        if (habitacion.getEstado() == EstadoHabitacion.OCUPADA) {
+            throw new HabitacionEnUsoException("No se puede eliminar una habitación ocupada");
+        }
+
+        //Falta validar que no haya reservas en la habitacion que se quiere borrar!!
 
         habitacionRepository.delete(habitacion);
     }
 
+    //5. Actualizaciones
+    // 5.1. Actualizar habitacion (completo)
     @Override
     public HabitacionDTOResponse update(Long id, HabitacionUpdateDTO dto) {
 
@@ -74,18 +111,23 @@ public class HabitacionServiceImpl implements HabitacionService {
         return habitacionMapper.toResponse(actualizada);
     }
 
-    //update
+    //5.2. Actualizar en repositorio
     @Override
     public void updateHabitacion(HabitacionEntity habitacionEntity){
         habitacionRepository.save(habitacionEntity);
     }
 
+    //5.3. Actualizar estado a MANTENIMIENTO
     @Override
     public HabitacionDTOResponse realizarMantenimiento(Long id) {
 
         HabitacionEntity habitacion = habitacionRepository.findById(id)
                 .orElseThrow(() ->
                         new HabitacionNoEncontradaException("Habitación no encontrada"));
+
+        if (habitacion.getEstado() == EstadoHabitacion.MANTENIMIENTO) {
+            throw new HabitacionYaEnMantenimientoException("La habitacion ya se encuentra en mantenimiento");
+        }
 
         habitacion.setEstado(EstadoHabitacion.MANTENIMIENTO);
 
@@ -95,6 +137,8 @@ public class HabitacionServiceImpl implements HabitacionService {
         return habitacionMapper.toResponse(actualizada);
     }
 
+    //6. Otras busquedas y listados
+    //6.1. Listar habitacioens disponibles
     @Override
     public List<HabitacionDTOResponse> habitacionesDisponibles() {
 
@@ -105,24 +149,14 @@ public class HabitacionServiceImpl implements HabitacionService {
                 .toList();
     }
 
-    @Override
-    public HabitacionEntity findEntityById(Long id) {
-
-        return habitacionRepository.findById(id)
-                .orElseThrow(() ->
-                        new HabitacionNoEncontradaException("Habitación no encontrada"));
-    }
-
+    //6.2. Listar habitaciones por estado
     @Override
     public List<HabitacionEntity> findByEstadoHabitacion(EstadoHabitacion estadoHabitacion) {
         return habitacionRepository.findByEstado(estadoHabitacion);
     }
 
-    @Override
-    public List<HabitacionEntity> findAll() {
-        return habitacionRepository.findAll();
-    }
-
+    //7. Calculos
+    //7.1. Determinar cantidad de habitaciones
     @Override
     public Integer cantidadHabitaciones() {
         return habitacionRepository.findAll().size();
